자신의 정보만 삭제 가능

+++ backend/src/main/java/com/example/backend/controller/member/MemberController.java
@@ -92,15 +92,22 @@ public class MemberController {
     }
 
     @DeleteMapping("remove")
-    public ResponseEntity<Map<String, Object>> remove(@RequestBody Member member) {
-        if (service.remove(member)) {
-            // 성공
-            return ResponseEntity.ok(Map.of("message", Map.of(
-                    "type", "success", "text", "회원 정보를 삭제하였습니다")));
+    @PreAuthorize("isAuthenticated()")
+    public ResponseEntity<Map<String, Object>> remove(
+            @RequestBody Member member,
+            Authentication authentication) {
+        if (service.hasAccess(member.getId(), authentication)) {
+            if (service.remove(member)) {
+                // 성공
+                return ResponseEntity.ok(Map.of("message", Map.of(
+                        "type", "success", "text", "회원 정보를 삭제하였습니다")));
+            } else {
+                // 실패
+                return ResponseEntity.badRequest().body(Map.of("message", Map.of(
+                        "type", "warning", "text", "정확한 정보를 입력해주세요")));
+            }
         } else {
-            // 실패
-            return ResponseEntity.badRequest().body(Map.of("message", Map.of(
-                    "type", "warning", "text", "정확한 정보를 입력해주세요")));
+            return ResponseEntity.status(403).build();
         }
     }
